Integrate Center admin with previous CenterAdmin product list
