+++ src/components/admin/AdminCenters.css
@@ -1,37 +1,3 @@
-*{
-	box-sizing: border-box;
-}
-a{
-	color: #33C3F0;
-}
-.container {
-	width: 80%;
-	margin: 0 auto;
-}
-
-.u-fl{
-	float: left;
-}
-.u-fr{
-	float: right;
-}
-.cf:before,
-.cf:after {
-    content: " "; /* 1 */
-    display: table; /* 2 */
-}
-
-.cf:after {
-    clear: both;
-}
-
-/**
- * For IE 6/7 only
- * Include this rule to trigger hasLayout and contain floats.
- */
-.cf {
-    *zoom: 1;
-}
 .admin-nav ul {
 	margin: 0;
 	padding: 0;
@@ -49,4 +15,3 @@ a{
 .edit-btn{
 	margin-right: 10px;
 }
-
