+++ src/components/admin/ViewCenter.css
@@ -1,0 +1,2 @@
+.ViewCenter {
+}
