--- src/screens/CenterAdmin.css
@@ -1,4 +1,0 @@
-
-.CenterAdmin {
-    text-align: center;
-}
